+++ app/components/views/TicketsPage/PurchaseTab/PurchaseTickets/index.js
@@ -8,7 +8,7 @@ import PurchaseTicketsAdvanced from "./PurchaseTicketsAdvanced";
 import PurchaseTicketsQuickBar from "./PurchaseTicketsQuickBar";
 import { injectIntl } from "react-intl";
 import { isNullOrUndefined } from "util";
-import { MIN_RELAY_FEE, MAX_POSSIBLE_FEE_INPUT } from "main_dev/constants";
+import { MIN_RELAY_FEE, MAX_POSSIBLE_FEE_INPUT } from "constants/Decred";
 
 @autobind
 class PurchaseTickets extends React.Component {
