+++ app/constants/Decred.js
@@ -1,0 +1,3 @@
+export const MAX_POSSIBLE_FEE_INPUT = 0.1;
+
+export const MIN_RELAY_FEE = 0.0001;
