+++ app/main_dev/paths.js
@@ -1,7 +1,7 @@
 import path from "path";
 import os from "os";
 import fs from "fs-extra";
-import { initWalletCfg, newWalletConfigCreation } from "../config";
+import { initWalletCfg, newWalletConfigCreation } from "config";
 
 // In all the functions below the Windows path is constructed based on
 // os.homedir() rather than using process.env.LOCALAPPDATA because in my tests
